edit generate annual report

+++ FinancialPlanningBE/FinancialPlanning.Data/Entities/ExpenseAnnualReport.cs
@@ -9,8 +9,8 @@ namespace FinancialPlanning.Data.Entities
     public class ExpenseAnnualReport
     {
         public string Department { get; set; }
-        public long TotalExpense { get; set; }
-        public long BiggestExpenditure { get; set; }
+        public decimal TotalExpense { get; set; }
+        public decimal BiggestExpenditure { get; set; }
         public string CostType { get; set; }
       
     }
